# docs/FE-contract/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "HistoryTalk API",
    "description": "API Specification for HistoryTalk project, synced with FE and latest DB specifications.",
    "version": "2.1"
  },
  "servers": [
    {
      "url": "http://localhost:8080/api/v1",
      "description": "Development Server"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/quizzes": {
      "get": {
        "tags": ["Quiz - Customer"],
        "summary": "Lấy danh sách bộ Quiz cho học sinh",
        "parameters": [
          { "name": "search", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Thành công",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ResponseWrapper" },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": { "$ref": "#/components/schemas/QuizSet" }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/quizzes/{quizId}": {
      "get": {
        "tags": ["Quiz - Customer"],
        "summary": "Chi tiết bộ Quiz",
        "parameters": [
          { "name": "quizId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Thành công",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ResponseWrapper" },
                    {
                      "type": "object",
                      "properties": {
                        "data": { "$ref": "#/components/schemas/QuizSet" }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/quizzes/{quizId}/start": {
      "post": {
        "tags": ["Quiz - Customer"],
        "summary": "Bắt đầu phiên làm bài Quiz",
        "parameters": [
          { "name": "quizId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Thành công",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ResponseWrapper" },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "sessionId": { "type": "string" },
                            "quizId": { "type": "string" },
                            "title": { "type": "string" },
                            "questions": {
                              "type": "array",
                              "items": { "$ref": "#/components/schemas/QuizQuestion" }
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/quizzes/submit": {
      "post": {
        "tags": ["Quiz - Customer"],
        "summary": "Nộp bài Quiz",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "sessionId": { "type": "string" },
                  "answers": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "questionId": { "type": "string" },
                        "selectedAnswer": { "type": "integer" }
                      }
                    }
                  }
                },
                "required": ["sessionId", "answers"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Thành công",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ResponseWrapper" },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "resultId": { "type": "string" },
                            "score": { "type": "number" },
                            "totalQuestions": { "type": "integer" },
                            "percentage": { "type": "number" },
                            "correctAnswers": { "type": "array", "items": { "type": "integer" } },
                            "wrongAnswers": { "type": "array", "items": { "type": "integer" } }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/quizzes/results/me": {
      "get": {
        "tags": ["Quiz - Customer"],
        "summary": "Xem lại các kết quả đã làm",
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 0 } },
          { "name": "size", "in": "query", "schema": { "type": "integer", "default": 10 } }
        ],
        "responses": {
          "200": {
            "description": "Thành công",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ResponseWrapper" },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "allOf": [
                            { "$ref": "#/components/schemas/Pagination" },
                            {
                              "type": "object",
                              "properties": {
                                "content": {
                                  "type": "array",
                                  "items": {
                                    "type": "object",
                                    "properties": {
                                      "resultId": { "type": "string" },
                                      "quizId": { "type": "string" },
                                      "quizTitle": { "type": "string" },
                                      "score": { "type": "number" },
                                      "totalQuestions": { "type": "integer" },
                                      "percentage": { "type": "number" },
                                      "completedAt": { "type": "string" }
                                    }
                                  }
                                }
                              }
                            }
                          ]
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Era": {
        "type": "string",
        "enum": ["ANCIENT", "MEDIEVAL", "MODERN", "CONTEMPORARY"]
      },
      "ResponseWrapper": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "message": { "type": "string" },
          "data": { "type": "object" },
          "timestamp": { "type": "string", "format": "date-time" }
        },
        "required": ["success", "message", "timestamp"]
      },
      "Pagination": {
        "type": "object",
        "properties": {
          "totalElements": { "type": "integer" },
          "totalPages": { "type": "integer" },
          "currentPage": { "type": "integer" },
          "pageSize": { "type": "integer" },
          "hasNext": { "type": "boolean" },
          "hasPrevious": { "type": "boolean" }
        },
        "required": ["totalElements", "totalPages", "currentPage", "pageSize", "hasNext", "hasPrevious"]
      },
...
      "QuizSet": {
        "type": "object",
        "properties": {
          "quizId": { "type": "string" },
          "title": { "type": "string" },
          "era": { "$ref": "#/components/schemas/Era" },
          "playCount": { "type": "integer" },
          "rating": { "type": "number" },
          "contextTitle": { "type": "string" }
        },
        "required": ["quizId", "title", "era", "playCount", "rating"]
      },
      "QuizQuestion": {
        "type": "object",
        "properties": {
          "questionId": { "type": "string" },
          "content": { "type": "string" },
          "options": {
            "type": "array",
            "items": { "type": "string" },
            "minItems": 4,
            "maxItems": 4
          },
          "correctAnswer": { "type": "integer" },
          "explanation": { "type": "string" }
        },
        "required": ["questionId", "content", "options", "correctAnswer"]
      },
...
    }
  }
}